got my sql graph to work
